migrate:role,user factory:articlecomment seeder:articles,role,user gate policy

+++ app/Http/Controllers/ArticleController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Articles;
 use App\Models\ArticleComment;
+use Illuminate\Support\Facades\Gate;
 
 class ArticleController extends Controller
 {
@@ -25,7 +26,14 @@ class ArticleController extends Controller
     }
 
     public function create(){
-        return view('articles.create');
+        $response = Gate::inspect('create');
+        if ($response->allowed()) {
+            return view('articles.create');
+        } else {
+            $response->message('1234');
+            var_dump($response);
+        }
+        
     }
 
     public function show($id){
@@ -46,11 +54,13 @@ class ArticleController extends Controller
     }
 
     public function update($id){
+        Gate::authorize('update-article');
         $article = Articles::findOrFail($id);
         return view('articles.edit', ['article' => $article]);
     }
 
     public function destroy($id){
+        Gate::authorize('delete-article');
         Articles::findOrFail($id)->delete();
         return redirect('/articles');
     }
